refactor: break the functions in small pieces
